+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## UNRELEASED - tbc
+
+### Removed
+
+- Support for capitalized grad-meh map paths, as it caused issues
+
+
 ## [1.0.0] - 2026-02-21
 
 ### Added